Create Practical2.py
Practical Title: A program that calculates the area and
perimeter of a rectangle.
Practical Description:
The objective of this practical is to develop a Python program
that efficiently calculates the area and perimeter of the
Rectangle. This exercise will help students understand basic
programming concepts like input/output operations, data types,
and arithmetic operations in Python.